Share test-changelog.xml and test.properties and verify.groovy
