+++ src/it/allChangesLiquibase411/verify.groovy
@@ -1,5 +1,3 @@
-import java.sql.ResultSet;
-
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -19,38 +17,5 @@ import java.sql.ResultSet;
  * under the License.
  */
 
-File buildLog = new File( basedir, 'build.log' )
-assert buildLog.exists()
-def buildLogText = buildLog.text;
-def defaultOptions = "--alter-foreign-keys-method=auto --no-check-unique-key-change --no-check-alter"
-assert buildLogText.contains("Executing: pt-online-schema-change ${defaultOptions} --alter=\"ADD COLUMN age INT NULL\" --password=*** --execute h=${config_host},P=${config_port},u=${config_user},D=testdb,t=person")
-assert buildLogText.contains("ChangeSet test-changelog.xml::2::Alice ran successfully")
-assert buildLogText.contains("Executing: pt-online-schema-change ${defaultOptions} --alter=\"DROP COLUMN age\"")
-assert buildLogText.contains("ChangeSet test-changelog.xml::3::Alice ran successfully")
-assert buildLogText.contains("Executing: pt-online-schema-change ${defaultOptions} --alter=\"ADD UNIQUE INDEX emailIdx (email)\"")
-assert buildLogText.contains("ChangeSet test-changelog.xml::4::Alice ran successfully")
-assert buildLogText.contains("Executing: pt-online-schema-change ${defaultOptions} --alter=\"DROP INDEX emailIdx\"")
-assert buildLogText.contains("ChangeSet test-changelog.xml::5::Alice ran successfully")
-assert buildLogText.contains("Executing: pt-online-schema-change ${defaultOptions} --alter=\"MODIFY email VARCHAR(400)\"")
-assert buildLogText.contains("ChangeSet test-changelog.xml::6::Alice ran successfully")
-assert buildLogText.contains("Executing: pt-online-schema-change ${defaultOptions} --alter=\"ADD CONSTRAINT fk_person_address FOREIGN KEY (person_id) REFERENCES person (name)\"")
-assert buildLogText.contains("ChangeSet test-changelog.xml::8::Alice ran successfully")
-assert buildLogText.contains("Executing: pt-online-schema-change ${defaultOptions} --alter=\"DROP FOREIGN KEY _fk_person_address\"")
-assert buildLogText.contains("ChangeSet test-changelog.xml::9::Alice ran successfully")
-assert buildLogText.contains("ChangeSet test-changelog.xml::10::Alice ran successfully")
-assert buildLogText.contains("Executing: pt-online-schema-change ${defaultOptions} --alter=\"DROP PRIMARY KEY, ADD PRIMARY KEY (id, name)\"")
-assert buildLogText.contains("ChangeSet test-changelog.xml::11::Alice ran successfully")
-
-File sql = new File( basedir, 'target/liquibase/migrate.sql' )
-assert sql.exists()
-def sqlText = sql.text;
-assert sqlText.contains("pt-online-schema-change ${defaultOptions} --alter=\"ADD COLUMN age INT NULL\"")
-assert sqlText.contains("pt-online-schema-change ${defaultOptions} --alter=\"DROP COLUMN age\"")
-assert sqlText.contains("pt-online-schema-change ${defaultOptions} --alter=\"ADD UNIQUE INDEX emailIdx (email)\"")
-assert sqlText.contains("pt-online-schema-change ${defaultOptions} --alter=\"DROP INDEX emailIdx\"")
-assert sqlText.contains("pt-online-schema-change ${defaultOptions} --alter=\"MODIFY email VARCHAR(400)\"")
-assert sqlText.contains("pt-online-schema-change ${defaultOptions} --alter=\"ADD CONSTRAINT fk_person_address FOREIGN KEY (person_id) REFERENCES person (name)\"")
-assert sqlText.contains("pt-online-schema-change ${defaultOptions} --alter=\"DROP FOREIGN KEY _fk_person_address\"")
-// Note: only adding the primary key, not dropping it in the migration sql.
-assert sqlText.contains("pt-online-schema-change ${defaultOptions} --alter=\"ADD PRIMARY KEY (id, name)\"")
-assert !sqlText.contains("password=${config_password}")
+File sharedVerifyScript = new File( basedir, '../sharedScripts/allChanges/verify.groovy' )
+evaluate( sharedVerifyScript )
