--- src/it/allChangesLiquibase443/test-changelog.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<databaseChangeLog
-        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.2.xsd">
-
-    <changeSet id="1" author="Alice">
-        <createTable tableName="person">
-            <column name="name" type="varchar(255)">
-                <constraints primaryKey="true"/>
-            </column>
-            <column name="email" type="varchar(255)"/>
-        </createTable>
-    </changeSet>
-
-    <changeSet id="2" author="Alice">
-        <addColumn tableName="person">
-            <column name="age" type="int"/>
-        </addColumn>
-    </changeSet>
-
-    <changeSet id="3" author="Alice">
-        <dropColumn tableName="person" columnName="age"/>
-    </changeSet>
-
-    <changeSet id="4" author="Alice">
-        <createIndex indexName="emailIdx" tableName="person" unique="true">
-            <column name="email"/>
-        </createIndex>
-    </changeSet>
-
-    <changeSet id="5" author="Alice">
-        <dropIndex indexName="emailIdx" tableName="person"/>
-    </changeSet>
-
-    <changeSet id="6" author="Alice">
-        <modifyDataType tableName="person" columnName="email" newDataType="VARCHAR(400)"/>
-    </changeSet>
-
-    <changeSet id="7" author="Alice">
-        <createTable tableName="address">
-            <column name="id" type="varchar(255)">
-                <constraints primaryKey="true"/>
-            </column>
-            <column name="person_id" type="varchar(255)"/>
-            <column name="street" type="varchar(255)"/>
-        </createTable>
-    </changeSet>
-
-    <changeSet id="8" author="Alice">
-        <addForeignKeyConstraint constraintName="fk_person_address"
-            referencedTableName="person" referencedColumnNames="name"
-            baseTableName="address" baseColumnNames="person_id"/>
-    </changeSet>
-
-    <changeSet id="9" author="Alice">
-        <dropForeignKeyConstraint baseTableName="address" constraintName="fk_person_address" />
-    </changeSet>
-
-    <changeSet id="10" author="Alice">
-        <createTable tableName="other_person">
-            <column name="id" type="varchar(255)">
-                <constraints primaryKey="true"/>
-            </column>
-            <column name="name" type="varchar(255)" />
-            <column name="email" type="varchar(255)"/>
-        </createTable>
-    </changeSet>
-
-    <changeSet id="11" author="Alice">
-        <addPrimaryKey tableName="other_person" columnNames="id, name"/>
-    </changeSet>
-</databaseChangeLog>
