+++ src/it/sharedScripts/allChanges/test.properties
